Fix copy code and fix layout for preview actions.
