+++ src/components/styled.js
@@ -1,4 +1,5 @@
 
+import {motion} from 'framer-motion';
 import styled from 'styled-components';
 
 export const CardTitle = styled.h4`
@@ -9,9 +10,33 @@ export const CardTitle = styled.h4`
 `;
 
 export const ExportCard = styled.div`
-    margin: 8px;
-    padding: 4px;
-    background: var(--back-light);
+    display: flex;
+    flex-direction:row-reverse;
+    padding: 0;
+    margin: 0;
+    width: auto;
+    align-content: flex-end;
+    z-index:99;
+    background: transparent;
     color: var(--prime);
 `;
 
+export const PreviewControlsDiv = styled(motion.div)`
+    display: flex;
+    flex-direction:row;
+    padding: 0;
+    margin: 0;
+    width: auto;
+    align-self: flex-start;
+    align-content: 
+    border: 1px solid var(--prime);
+    z-index:99;
+`;
+
+export const PreviewControlsDirectionals = styled(motion.div)`
+    display: flex;
+    flex-direction: column;
+    align-content: center;
+    font-size: 20pt;
+    padding: 0;
+`
